+++ zellij-utils/src/errors.rs
@@ -436,6 +436,7 @@ pub enum ScreenContext {
     ToggleGroupMarking,
     SessionSharingStatusChange,
     SetMouseSelectionSupport,
+    SetPaneCollapsed,
     InterceptKeyPresses,
     ClearKeyPressesIntercepts,
     ReplacePaneWithExistingPane,